start of 2/23/22

+++ Klondike-master/src/Location.java
@@ -1,13 +1,13 @@
 public class Location {
-    private int x, y;
-    public Location(int x, int y){
+    private double x, y;
+    public Location(double x, double y){
         this.x=x;
         this.y=y;
     }
-    public int getX(){
+    public double getX(){
         return x;
     }
-    public int getY(){
+    public double getY(){
         return y;
     }
 }
